replace index.d.ts with index.ts
